# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Secure E-Wallet API
  version: 0.2.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200": { description: OK }
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      summary: Logout (invalidate token)
      responses:
        "204": { description: Logged out }
  /wallet/me:
    get:
      summary: Get current user's wallet
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/deposit:
    post:
      summary: Deposit funds
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /transfer:
    post:
      summary: Transfer funds to another user
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "200":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"
  /transactions:
    get:
      summary: List transactions
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
  /security/login-events:
    post:
      summary: Record login event and request anomaly score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginEvent"
      responses:
        "200":
          description: AI score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyScore"
  /security/alerts:
    get:
      summary: List security alerts for current user
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SecurityAlert" }
  /admin/users:
    get:
      summary: Admin list users
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
  /admin/alerts:
    get:
      summary: Admin list alerts
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SecurityAlert" }
  /admin/audit-logs:
    get:
      summary: Admin list audit logs
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditLog" }
  /admin/policies:
    get:
      summary: Admin get security policies
      security: [ { bearerAuth: [] } ]
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SecurityPolicy" }
    post:
      summary: Admin update policy
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecurityPolicy"
      responses:
        "200": { description: Updated }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [USER, ADMIN] }
    Wallet:
      type: object
      required: [id, balance, currency]
      properties:
        id: { type: string }
        balance: { type: number, format: double }
        currency: { type: string, default: USD }
    Transaction:
      type: object
      required: [id, amount, type, createdAt]
      properties:
        id: { type: string }
        amount: { type: number, format: double }
        type: { type: string }
        description: { type: string }
        createdAt: { type: string, format: date-time }
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [USER, ADMIN] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    DepositRequest:
      type: object
      required: [amount]
      properties:
        amount: { type: number, format: double, minimum: 0 }
    TransferRequest:
      type: object
      required: [toUserId, amount]
      properties:
        toUserId: { type: string }
        amount: { type: number, format: double, minimum: 0 }
    TransferResult:
      type: object
      properties:
        status: { type: string }
        transaction: { $ref: "#/components/schemas/Transaction" }
    LoginEvent:
      type: object
      required: [userId, ipAddress, timestamp]
      properties:
        userId: { type: string }
        ipAddress: { type: string }
        userAgent: { type: string }
        timestamp: { type: string, format: date-time }
    AnomalyScore:
      type: object
      properties:
        score: { type: number, format: double }
        reasons:
          type: array
          items: { type: string }
    SecurityAlert:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    AuditLog:
      type: object
      properties:
        id: { type: string }
        action: { type: string }
        actor: { type: string }
        details: { type: string }
        createdAt: { type: string, format: date-time }
    SecurityPolicy:
      type: object
      properties:
        id: { type: string }
        maxLoginAttempts: { type: integer }
        lockoutMinutes: { type: integer }
        rateLimitPerMin: { type: integer }